[funding]refactor:금액 타입 변경 #32

+++ src/main/java/com/pieceofcake/fundingservice/funding/dto/in/CreateFundingRequestDto.java
@@ -20,8 +20,8 @@ import java.util.UUID;
 public class CreateFundingRequestDto {
     private String fundingUuid;
     private String productUuid;
-    private Double fundingAmount;
-    private Double piecePrice;
+    private Long fundingAmount;
+    private Long piecePrice;
     private Integer totalPieces;
     private Integer remainingPieces;
     private LocalDateTime fundingDeadline;
